Design Resources!  Any Module file can be overridden in the repository.  Structures save as special Design Resources.  Indexes are configured in Modules and stored as temporary files.
git-svn-id: https://svn.apache.org/repos/asf/lenya/branches/revolution@693670 13f79535-47bb-0310-9956-ffa450edef68

+++ 1.3.x/src/java/org/apache/lenya/cms/content/flat/FlatTranslation.java
@@ -1,7 +1,11 @@
 package org.apache.lenya.cms.content.flat;
 import java.io.File;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
 import java.util.SortedSet;
 import java.util.TreeSet;
+import org.apache.lenya.cms.content.Content;
 import org.apache.lenya.util.Globals;
 import org.apache.lenya.xml.DocumentHelper;
 import org.w3c.dom.Document;
@@ -12,54 +16,52 @@ import org.w3c.dom.Element;
  * @since 1.3
  */
 public class FlatTranslation {
-   File resourceDirectory;
-   File translationDirectory;
-   String language = "en";
-   String defaultLanguage = "en";
-   String live = "";
-   String edit = "last";
-   String revision = "live";
-   Document document;
-   Element root;
-   FlatRevision fr;
-   boolean initRevisions = false;
-   SortedSet revisions = new TreeSet();
-   boolean isChanged = false;
-   public FlatTranslation(File directory, String planguage, String prevision, String pdefaultLanguage) {
+   private File resourceDirectory;
+   private File translationDirectory;
+   private String language = "en";
+   private String defaultLanguage = "en";
+   private String revision = Content.REVISION_DEFAULT;
+   private Document document;
+   private Element root;
+   private FlatRevision flatRevision;
+   private boolean initRevisions = false;
+   private SortedSet revisions = new TreeSet(); // list of FlatRevisions' filenames without extension
+   private Map namedRevisions = new HashMap();
+   private boolean isChanged = false;
+   public FlatTranslation(File directory, String language, String revision, String defaultLanguage) {
+      // System.out.println("FlatTranslation new L=" + language + " DL=" + defaultLanguage + " REV=" + revision + " DIR=" + directory.getAbsolutePath());
       resourceDirectory = directory;
-      language = planguage;
-      defaultLanguage = pdefaultLanguage;
-      revision = prevision;
+      this.language = language;
+      this.defaultLanguage = defaultLanguage;
+      this.revision = revision;
       init();
    }
-   public FlatTranslation(File directory, String planguage, String prevision) {
+   public FlatTranslation(File directory, String language, String revision) {
+      // System.out.println("FlatTranslation new L=" + language + " REV=" + revision + " DIR=" + directory.getAbsolutePath());
       resourceDirectory = directory;
-      language = planguage;
-      revision = prevision;
-      defaultLanguage = planguage; // Set to publication's default
+      this.language = language;
+      this.revision = revision;
+      this.defaultLanguage = language; // Set to publication's default
       init();
    }
-   public FlatTranslation(File directory, String planguage) {
+   public FlatTranslation(File directory, String language) {
+      // System.out.println("FlatTranslation new L=" + language + " DIR=" + directory.getAbsolutePath());
       resourceDirectory = directory;
-      language = planguage;
-      defaultLanguage = planguage; // Set to publication's default
+      this.language = language;
+      this.defaultLanguage = language; // Set to publication's default
       init();
    }
    private void init() {
       translationDirectory = new File(resourceDirectory, language);
-      if(!translationDirectory.exists())
-         translationDirectory = new File(resourceDirectory, defaultLanguage);
-      if(!translationDirectory.exists())
-         return;
+      if(!translationDirectory.exists()) translationDirectory = new File(resourceDirectory, defaultLanguage);
+      if(!translationDirectory.exists()) return;
       File translationFile = new File(translationDirectory, "translation.xml");
       if(translationFile.exists()){
          try{
             document = DocumentHelper.readDocument(translationFile);
             root = document.getDocumentElement();
-            if(root.hasAttribute("live"))
-               live = root.getAttribute("live");
-            if(root.hasAttribute("edit"))
-               edit = root.getAttribute("edit");
+            if(root.hasAttribute(Content.REVISION_DEFAULT)) setNamedRevision(Content.REVISION_DEFAULT, root.getAttribute(Content.REVISION_DEFAULT));
+            if(root.hasAttribute("edit")) setNamedRevision("edit", root.getAttribute("edit"));
          }catch(javax.xml.parsers.ParserConfigurationException pce){
             System.out.println("FlatTranslation - ParserConfigurationException: " + translationFile.getAbsolutePath());
          }catch(org.xml.sax.SAXException saxe){
@@ -82,102 +84,143 @@ public class FlatTranslation {
    public String[] getRevisions() {
       if(!initRevisions){
          String[] filelist = translationDirectory.list();
-         for(int f = 0; f < filelist.length; f++){
-            String filename = filelist[f];
-            int pos = filename.lastIndexOf(".");
-            if(pos > 0)
-               filename = filename.substring(0, pos);
-            if(!filename.equalsIgnoreCase("translation"))
-               revisions.add(filename);
+         if(null == filelist){
+            // System.out.println("FlatTranslatioOn.getRevisions NULL=" + translationDirectory.getAbsolutePath());
+         }else{
+            for(int f = 0; f < filelist.length; f++){
+               String filename = filelist[f];
+               int pos = filename.lastIndexOf(".");
+               if(pos > 0) filename = filename.substring(0, pos);
+               if(!filename.equalsIgnoreCase("translation")) revisions.add(filename);
+            }
+            initRevisions = true;
          }
-         initRevisions = true;
       }
       return (String[]) revisions.toArray(new String[0]);
    }
-   public String getLive() {
-      return live;
-   }
-   public String getEdit() {
-      return edit;
-   }
    public String getURI() {
-      if(null == fr)
-         fr = getRevision();
-      if(null == fr){
-         // System.out.println("FlatTranslation.getURI no revision.");
+      if(null == flatRevision) flatRevision = getRevision();
+      if(null == flatRevision){
+         // System.out.println("FlatTranslation.getURI: No revision.");
          return "";
       }
-      return fr.getURI();
+      return flatRevision.getURI();
    }
    public String getMetaURI() {
-      if(null == fr)
-         fr = getRevision();
-      if(null == fr)
-         return "";
-      return fr.getMetaURI();
+      if(null == flatRevision) flatRevision = getRevision();
+      if(null == flatRevision) return "";
+      return flatRevision.getMetaURI();
    }
    public String getNewURI() {
       String newRevision = Globals.getDateString();
       // WORK: Change Edit to newRevision
-      return new File(translationDirectory, newRevision + ".xml").getPath();
+      return (new File(translationDirectory, newRevision + ".xml")).getPath();
    }
    public String getExtension() {
-      if(null == fr)
-         fr = getRevision();
-      if(null == fr)
-         return "";
-      return fr.getExtension();
+      if(null == flatRevision) flatRevision = getRevision();
+      if(null == flatRevision) return "";
+      return flatRevision.getExtension();
    }
    public String getHREF() {
-      if(null == fr)
-         fr = getRevision();
-      if(null == fr)
-         return "";
-      return fr.getHREF();
+      if(null == flatRevision) flatRevision = getRevision();
+      if(null == flatRevision) return "";
+      return flatRevision.getHREF();
    }
    public FlatRevision getRevision() {
       return getRevision(revision);
    }
-   public FlatRevision getRevision(String prevision) {
-      String rev = prevision;
-      if(rev.equalsIgnoreCase("edit")){
-         if(edit.length() > 0){
-            rev = edit;
-         }else
-            rev = live;
+   public FlatRevision getRevision(String revision) {
+      // System.out.println("FlatTranslation.getRevision " + revision);
+      String rev = revision;
+      // Verify revisions is initialized.
+      getRevisions();
+      if(revisions.isEmpty()){
+         rev = Globals.getDateString();
+         initRevisions = false;
       }
-      if(rev.equalsIgnoreCase("live"))
-         rev = live;
+      if(rev.equalsIgnoreCase("edit")) rev = getNamedRevision("edit");
+      if(rev.equalsIgnoreCase(Content.REVISION_DEFAULT)) rev = getNamedRevision(Content.REVISION_DEFAULT);
       try{
-         if(rev.equalsIgnoreCase("last"))
-            rev = (String) revisions.last();
-         if(rev.equalsIgnoreCase("first"))
-            rev = (String) revisions.first();
+         if(rev.equalsIgnoreCase("last")) rev = (String) revisions.last();
+         if(rev.equalsIgnoreCase("first")) rev = (String) revisions.first();
       }catch(java.util.NoSuchElementException nsee){
          return null;
       }
-      if(null == rev)
+      // Defaults
+      // System.out.println("FlatTranslation.getRevision rev1=" + rev);
+      if(1 > rev.length()) rev = getNamedRevision(Content.REVISION_DEFAULT);
+      try{
+         if(1 > rev.length()) rev = (String) revisions.last();
+      }catch(java.util.NoSuchElementException nsee){
          return null;
-      if(rev.length() < 1)
+      }
+      if((null == rev) || (1 > rev.length())){
+         // System.out.println("FlatTranslation.getRevision rev=null");
          return null;
+      }
+      // System.out.println("FlatTranslation.getRevision revZ=" + rev);
       return new FlatRevision(translationDirectory, rev);
    }
-   void setLive(String revision) {
-      live = revision;
+   void setNamedRevision(String name, String revision) {
+      namedRevisions.put(name, revision);
       isChanged = true;
    }
+   String getNamedRevision(String name) {
+      String revision = "";
+      if(namedRevisions.containsKey(name)){
+         revision = (String) namedRevisions.get(name);
+      }
+      if((1 > revision.length()) && namedRevisions.containsKey(Content.REVISION_DEFAULT)){
+         revision = (String) namedRevisions.get(Content.REVISION_DEFAULT);
+      }
+      if((1 > revision.length()) && namedRevisions.containsKey("last")) revision = (String) namedRevisions.get("last");
+      return revision;
+   }
+   /**
+    * @deprecated Use getNamedRevision(Content.REVISION_DEFAULT)
+    * @return
+    */
+   public String getLive() {
+      return getNamedRevision(Content.REVISION_DEFAULT);
+   }
+   /**
+    * @deprecated Use getNamedRevision("edit")
+    * @return
+    */
+   public String getEdit() {
+      return getNamedRevision("edit");
+   }
+   /**
+    * @deprecated Use setNamedRevision(Content.REVISION_DEFAULT, revision)
+    * @param revision
+    */
+   void setLive(String revision) {
+      setNamedRevision(Content.REVISION_DEFAULT, revision);
+      isChanged = true;
+   }
+   /**
+    * @deprecated Use setNamedRevision("edit", revision)
+    * @param revision
+    */
    void setEdit(String revision) {
-      edit = revision;
+      setNamedRevision("edit", revision);
       isChanged = true;
    }
    void save() {
       if(isChanged){
+         FlatContent content = (FlatContent) Globals.getPublication().getContent();
          File file = new File(translationDirectory, "translation.xml");
-         /*
-          * Document doc; try{ doc = DocumentHelper.readDocument(file); }catch(javax.xml.parsers.ParserConfigurationException pce){ System.out.println("FlatTranslation.save - ParserConfigurationException:" + file.getAbsolutePath()); return; }catch(org.xml.sax.SAXException saxe){ System.out.println("FlatTranslation.save - SAXException:" + file.getAbsolutePath()); return; }catch(java.io.IOException ioe){ System.out.println("FlatTranslation.save - Could not read file:" + file.getAbsolutePath()); return; } root = doc.getDocumentElement();
-          */
-         root.setAttribute("live", live);
-         root.setAttribute("edit", edit);
+         // Used NamedRevisions
+         Iterator revisionNames = content.getRevisions().iterator();
+         while(revisionNames.hasNext()){
+            String revisionName = (String) revisionNames.next();
+            // TODO: NPE on setAttribute
+            // System.out.println("FlatTranslation.save - RevName:" + revisionName + " Rev:" + getNamedRevision(revisionName));
+            String revision = getNamedRevision(revisionName);
+            // if(0 < revision.length())
+            // Zero-length revision is valid for retracting (unpublishing).
+            root.setAttribute(revisionName, revision);
+         }
          try{
             // DocumentHelper.writeDocument(doc, file);
             DocumentHelper.writeDocument(document, file);
@@ -188,7 +231,8 @@ public class FlatTranslation {
          }catch(java.io.IOException ioe2){
             System.out.println("FlatTranslation.save - Could not write file:" + file.getAbsolutePath());
          }
-         // TODO: Update indexes for this language.
+         // TODO: Update indexes for just this language.
+         content.updateIndexes();
       }
       isChanged = false;
    }
@@ -201,9 +245,11 @@ public class FlatTranslation {
          file.delete();
          if(extension.length() > 0){
             file = new File(translationDirectory, revision + "." + extension);
-            if(file.exists())
-               file.delete();
+            if(file.exists()) file.delete();
          }
       }
+   }
+   public boolean exists() {
+      return getRevision().exists();
    }
 }
